Add attribute clustering notebook
Adds the ipynb that i was using for the analysis. I am commiting it like this because to get it to work you have to change a lot of the imports in the core folder, however, if I changed these they would break the app. So, instead, just downloaded the notebook and am uploading it like this.